Alarm and show alarm implemented

+++ app/src/main/java/android/bignerdranch/bateopontoai/data/HoursAndMinutes.kt
@@ -1,6 +1,6 @@
 package android.bignerdranch.bateopontoai.data
 
 data class HoursAndMinutes(
-    val hours : Int? = null,
-    val minutes : Int? = null
+    val hours: Long? = null,
+    val minutes: Int? = null
 )
